Unit test addition for main form closing
Added a unit test to ensure that the application main form is closed when
is should be. Since raising and consuming the FormClosing event would be
cumbersome, the event consumer method is called directly from the test
code. This required adding an interface for the main form. The interface
will be handy for other event related tests later on.