Last time period not updated correctly
The current, non-completed time period duration was only correct when the
time report was shown for the first time; it was not updated as time
passed.

Added a unit test to detect the bug and fixed it by updating the duration
in TimeTracker#GetPeriods. The time period duration is now mutable.